+++ src/services/MusicService.ts
@@ -20,6 +20,7 @@ import { IsLikedInfoRDB } from '../interfaces/user/IsLikedInfoRDB';
 import { MumentAndUserInfoRDB } from '../interfaces/mument/MumentAndUserInfoRDB';
 import common from '../modules/common';
 import { TagListInfo } from '../interfaces/common/TagListInfo';
+import appleMusic from '../library/appleMusicSearch';
 
 const qs = require('querystring');
 require('dotenv').config();
@@ -319,62 +320,65 @@ const getMumentList = async (musicId: string, userId: string, isLikeOrder: boole
 };
 
 /**
- * 곡 검색 - apple music api 사용 곡 검색 / 최대 25개의 곡 리스트 반환 가능
+ * 곡 검색 - apple music api 사용 곡 검색 / 최대 50개까지 곡 리스트 반환 가능
  */
 const getMusicListBySearch = async (keyword: string): Promise<MusicResponseDto[] | number | void> => {
     try {        
-        const token = `Bearer ${config.appleDeveloperToken as string}`;
+        //const token = `Bearer ${config.appleDeveloperToken as string}`;
 
-        let musicList: MusicResponseDto[] = [];
-
-        const appleResponse = async (searchKeyword: string) => {
+        // const appleResponse = async (searchKeyword: string) => {
     
-            await axios.get('https://api.music.apple.com/v1/catalog/kr/search?types=songs&limit=20&term=' 
-                + encodeURI(searchKeyword), {
-                    headers: {
-                      'Content-Type': 'application/x-www-form-urlencoded',
-                      'Authorization': token
-                    }
-                }
-            )
-            .then(async function (response: any) {
-                /* apple api에서 받을 수 있는 3개 status code 대응 - 200, 401, 500*/       
+        //     await axios.get(`https://api.music.apple.com/v1/catalog/kr/search?types=songs&limit=25&offset=${0}&term=`
+        //         + encodeURI(searchKeyword), {
+        //             headers: {
+        //               'Content-Type': 'application/x-www-form-urlencoded',
+        //               'Authorization': token
+        //             }
+        //         }
+        //     )
+        //     .then(async function (response: any) {
+        //         /* apple api에서 받을 수 있는 3개 status code 대응 - 200, 401, 500*/       
 
-                if (response.data.results.hasOwnProperty('songs')) {
-                    // 401 - A response indicating an incorrect Authorization header
-                    if (response.status == 401) return constant.APPLE_UNAUTHORIZED;
+        //         if (response.data.results.hasOwnProperty('songs')) {
+        //             // 401 - A response indicating an incorrect Authorization header
+        //             if (response.status == 401) return constant.APPLE_UNAUTHORIZED;
 
-                    // 500 - indicating an error occurred on the apple music server
-                    if (response.status == 500) return constant.APPLE_INTERNAL_SERVER_ERROR;
+        //             // 500 - indicating an error occurred on the apple music server
+        //             if (response.status == 500) return constant.APPLE_INTERNAL_SERVER_ERROR;
 
-                    const appleMusicList = response.data.results.songs.data; 
+        //             const appleMusicList = response.data.results.songs.data; 
 
-                    musicList =  await appleMusicList.map((music: any) => {
-                        let imageUrl = music.attributes.artwork.url;
-                        imageUrl = imageUrl.replace('{w}x{h}', '400x400'); //앨범 이미지 크기 400으로 지정
+        //             musicList =  await appleMusicList.map((music: any) => {
+        //                 let imageUrl = music.attributes.artwork.url;
+        //                 imageUrl = imageUrl.replace('{w}x{h}', '400x400'); //앨범 이미지 크기 400으로 지정
 
-                        const result: MusicResponseDto = {
-                            '_id': music.id,
-                            'name': music.attributes.name,
-                            'artist': music.attributes.artistName,
-                            'image': imageUrl
-                        };
-                        return result;
-                    });
-                }
+        //                 const result: MusicResponseDto = {
+        //                     '_id': music.id,
+        //                     'name': music.attributes.name,
+        //                     'artist': music.attributes.artistName,
+        //                     'image': imageUrl
+        //                 };
+        //                 return result;
+        //             });
+        //         }
                 
-                return musicList;
-            })
-            .catch(async function (error) {
-                console.log('곡검색 애플 error', error);
-                return constant.APPLE_INTERNAL_SERVER_ERROR;
-            });
+        //         return musicList;
+        //     })
+        //     .catch(async function (error) {
+        //         console.log('곡검색 애플 error', error);
+        //         return constant.APPLE_INTERNAL_SERVER_ERROR;
+        //     });
 
-            return musicList;
-        };
-        const data = await appleResponse(keyword);
+        //     return musicList;
+        // };
+        
+        // 곡 검색 첫 페이지 개수가 25개 이상일 경우만 검색 2회 요청
+        const page1 = await appleMusic.searchMusic(keyword, 0);
+        if (page1.length < 25) return page1;
+        
+        const page2 = await appleMusic.searchMusic(keyword, 25);
 
-        return data;
+        return page1.concat(page2);
     } catch (error) {
         console.log(error);
         throw error;
